Add more OOP and package all classes into one package
