+++ src/main/java/duke/Deadline.java
@@ -1,3 +1,4 @@
+package duke;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
